requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
google-generativeai
pydantic[email]
python-multipart